trading-bot/requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
exchange-calendars>=4.5.0
aiohttp>=3.9.0
pytz>=2024.1
websockets>=12.0
